Add use case test for merging scala modules
